Separated the division functions into a separate file for unit testing
