+++ examples/asrc_demo/src/rate_server.c
@@ -27,6 +27,7 @@
 #include "rate_server.h"
 #include "avg_buffer_level.h"
 #include "tusb.h"
+#include "div.h"
 
 #define LOG_I2S_TO_USB_SIDE (0)
 #define LOG_USB_TO_I2S_SIDE (0)
@@ -308,90 +309,6 @@ void rate_server(void *args)
     }
 }
 
-#if __xcore__
-uint32_t dsp_math_divide_unsigned(uint32_t dividend, uint32_t divisor, uint32_t q_format )
-{
-    //h and l hold a 64-bit value
-    uint32_t h; uint32_t l;
-    uint32_t quotient=0, remainder=0;
-
-    // Create long dividend by shift dividend up q_format positions
-    h = dividend >> (32-q_format);
-    l = dividend << (q_format);
-
-    // Unsigned Long division
-    asm("ldivu %0,%1,%2,%3,%4":"=r"(quotient):"r"(remainder),"r"(h),"r"(l),"r"(divisor));
-
-    return quotient;
-}
-#else //__xcore__
-// If we're compiling this for x86 we're probably testing it - let the compiler work out the ASM for this
-
-uint32_t dsp_math_divide_unsigned(uint32_t dividend, uint32_t divisor, uint32_t q_format )
-{
-    uint64_t h = (uint64_t)dividend << q_format;
-    uint64_t quotient = h / divisor;
-
-    return (uint32_t)quotient;
-}
-#endif //__xcore__
-
-float_s32_t float_div(float_s32_t dividend, float_s32_t divisor)
-{
-    float_s32_t res;// = float_s32_div(dividend, divisor);
-
-    int dividend_hr;
-    int divisor_hr;
-
-    asm( "clz %0, %1" : "=r"(dividend_hr) : "r"(dividend.mant) );
-    asm( "clz %0, %1" : "=r"(divisor_hr) : "r"(divisor.mant) );
-
-
-    int dividend_exp = dividend.exp - dividend_hr;
-    int divisor_exp = divisor.exp - divisor_hr;
-
-
-    uint64_t h_dividend = (uint64_t)((uint32_t)dividend.mant) << (dividend_hr);
-
-    uint32_t h_divisor = ((uint32_t)divisor.mant) << (divisor_hr);
-
-    uint32_t lhs = (h_dividend > h_divisor) ? 31 : 32;
-
-    uint64_t normalised_dividend = h_dividend << lhs;
-
-#if __xcore__
-    uint32_t quotient = 0;
-    uint32_t remainder = 0;
-    uint32_t h = (uint32_t)(normalised_dividend>>32);
-    uint32_t l = (uint32_t)(normalised_dividend);
-    asm("ldivu %0,%1,%2,%3,%4":"=r"(quotient):"r"(remainder),"r"(h),"r"(l),"r"(h_divisor));
-#else
-    uint64_t quotient = (uint64_t)(normalised_dividend) / h_divisor;
-#endif
-
-    res.exp = dividend_exp - divisor_exp - lhs;
-
-    res.mant = (uint32_t)(quotient) ;
-    return res;
-}
-
-uint32_t float_div_fixed_output_q_format(float_s32_t dividend, float_s32_t divisor, int32_t output_q_format)
-{
-    int op_q = -output_q_format;
-    float_s32_t res = float_div(dividend, divisor);
-    uint32_t quotient;
-    if(res.exp < op_q)
-    {
-        int rsh = op_q - res.exp;
-        quotient = ((uint32_t)res.mant >> rsh) + (((uint32_t)res.mant >> (rsh-1)) & 0x1);
-    }
-    else
-    {
-        int lsh = res.exp - op_q;
-        quotient = (uint32_t)res.mant << lsh;
-    }
-    return quotient;
-}
 
 uint32_t sum_array(uint32_t * array_to_sum, uint32_t array_length)
 {
@@ -401,55 +318,4 @@ uint32_t sum_array(uint32_t * array_to_sum, uint32_t array_length)
         acc += array_to_sum[i];
     }
     return acc;
-}
-
-typedef struct
-{
-    uint64_t mant;
-    int32_t exp;
-}float_u64_t;
-
-static float_u64_t float_div_u64(float_s32_t dividend, float_s32_t divisor)
-{
-    float_u64_t res;
-
-    int dividend_hr;
-    int divisor_hr;
-
-    asm( "clz %0, %1" : "=r"(dividend_hr) : "r"(dividend.mant) );
-    asm( "clz %0, %1" : "=r"(divisor_hr) : "r"(divisor.mant) );
-
-    int dividend_exp = dividend.exp - dividend_hr;
-    int divisor_exp = divisor.exp - divisor_hr;
-
-    uint64_t h_dividend = (uint64_t)((uint32_t)dividend.mant) << (dividend_hr);
-
-    uint32_t h_divisor = ((uint32_t)divisor.mant) << (divisor_hr);
-
-    uint32_t lhs = 32;
-
-    uint64_t quotient = (h_dividend << lhs) / h_divisor;
-
-    res.exp = dividend_exp - divisor_exp - lhs;
-
-    res.mant = quotient ;
-    return res;
-}
-
-uint64_t float_div_u64_fixed_output_q_format(float_s32_t dividend, float_s32_t divisor, int32_t output_q_format)
-{
-    int op_q = -output_q_format;
-    float_u64_t res = float_div_u64(dividend, divisor);
-    uint64_t quotient;
-    if(res.exp < op_q)
-    {
-        int rsh = op_q - res.exp;
-        quotient = ((uint64_t)res.mant >> rsh) + (((uint64_t)res.mant >> (rsh-1)) & 0x1);
-    }
-    else
-    {
-        int lsh = res.exp - op_q;
-        quotient = (uint64_t)res.mant << lsh;
-    }
-    return quotient;
 }
