+++ examples/asrc_demo/src/rate_server.h
@@ -14,9 +14,6 @@ void set_spkr_itf_close_open_event(bool event);
 
 // Rate calculation math functions
 uint32_t sum_array(uint32_t * array_to_sum, uint32_t array_length);
-float_s32_t float_div(float_s32_t dividend, float_s32_t divisor);
-uint32_t float_div_fixed_output_q_format(float_s32_t dividend, float_s32_t divisor, int32_t output_q_format);
-uint64_t float_div_u64_fixed_output_q_format(float_s32_t dividend, float_s32_t divisor, int32_t output_q_format);
 
 // Wrapper functions for calculating i2s send buffer average level
 void init_calc_i2s_buffer_level_state(void);
